Crate zng_app_proc_macros

Proc-macros for zng-app.

Crate

This crate is part of the zng project.

Attribute Macros

easing

Expands a property assign to include a build action that applies an easing transition to the variable inputs.

property

Expands a function to a widget property.

widget

Expands a struct to a widget and macro.

widget_mixin

Expands a generic struct to a widget mixin.